LayoutTests/fast/animation/css-animation-throttling.html

<!DOCTYPE html> <!-- webkit-test-runner [ RespondToThermalPressureAggressively=true ] -->
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<style>
@keyframes my-animation {        
  from {
    left : 0px;
  }
  to {
    left : 100px;
  }      
}

.run-animation {
  position: relative;
  animation: my-animation 0.5s infinite;
  animation-direction: alternate;
}
</style>
</head>
<body>
<p id="testElement" class="run-animation">Test text</p>
<script>
description("Tests that CSS animations are throttled in low power mode.");
jsTestIsAsync = true;

document.getElementById("testElement").addEventListener("animationstart", () => {
    debug("Ensure low-power mode and aggressive thermal mitigation are disabled");
    evalAndLog("internals.setLowPowerModeEnabled(false)");
    evalAndLog("internals.setAggressiveThermalMitigationEnabled(false)");
    debug("");

    window.unthrottled = internals.animationsInterval;
    window.throttled = unthrottled * 2;

    evalAndLog("internals.setLowPowerModeEnabled(true)");
    shouldBe("internals.animationsInterval", "throttled");
    evalAndLog("internals.setLowPowerModeEnabled(false)");
    shouldBe("internals.animationsInterval", "unthrottled");
    evalAndLog("internals.setAggressiveThermalMitigationEnabled(true)");
    shouldBe("internals.animationsInterval", "throttled");
    evalAndLog("internals.setAggressiveThermalMitigationEnabled(false)");
    shouldBe("internals.animationsInterval", "unthrottled");
    finishJSTest();
}, { once: true });
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
