LayoutTests/fast/animation/css-animation-throttling-expected.txt

Tests that CSS animations are throttled in low power mode.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Ensure low-power mode and aggressive thermal mitigation are disabled
internals.setLowPowerModeEnabled(false)
internals.setAggressiveThermalMitigationEnabled(false)

internals.setLowPowerModeEnabled(true)
PASS internals.animationsInterval is throttled
internals.setLowPowerModeEnabled(false)
PASS internals.animationsInterval is unthrottled
internals.setAggressiveThermalMitigationEnabled(true)
PASS internals.animationsInterval is throttled
internals.setAggressiveThermalMitigationEnabled(false)
PASS internals.animationsInterval is unthrottled
PASS successfullyParsed is true

TEST COMPLETE
Test text
